+++ src/create-variable.test.ts
@@ -45,6 +45,18 @@ it("should create the variable just before usage (variable declaration)", () => 
   });
 });
 
+it("should create the variable just before usage (JSX attribute)", () => {
+  shouldUpdateCodeFor({
+    code: `function MyButton() {
+  return <button onClick={handleClick} />
+}`,
+    selection: Selection.cursorAt(1, 26),
+    expected: {
+      position: new Position(1, 2)
+    }
+  });
+});
+
 it("should respect code indentation", () => {
   shouldUpdateCodeFor({
     code: `function sayHello() {