+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- "Create Variable" now works well from variable declarations
+- "Create Variable" now works well from variable declarations & JSX syntax
 
 ## [1.2.0] - 2020-04-04 - I handle async functions 🔮